Add .gitignore files to preserve empty directories
The source release contained empty `linux/` directories, but these
were dropped when converted to Git, because Git doesn't track empty
directories.  This is unfortunate, because the Makefile relies on the
directory already existing.  Instead, add an empty `.gitignore` file
to each empty directory, so that Git recognizes the directory.

This brings the Git tree nearly in-line with the original 1997-12-23
id source release:

    ftp://ftp.idsoftware.com/idstuff/source/doomsrc.zip
    SHA256 (doomsrc.zip) = 29348bc1164e92b3e24807c6013db88b2847dd7fdf03f85261951e38de06a337

(While ftp.idsoftware.com is no longer around, there several mirrors
of it around--you should not have troule finding this doomsrc.zip.)

That original doomsrc.zip contains 3 files:

    README.TXT
    linuxdoom-1.10.src.tgz
    sndserv.tgz

If you extract the doomsrc.zip, then extract each of the .tgz files,
the difference between that and this Git tree are as follows:

 - doomsrc.zip's README.TXT uses CRLF newlines, while the Git version
   uses LF newlines.
 - The .tgz files are not checked in to Git, just their extracted
   contents are.
 - The Git version contains the two additional .gitignore files to
   preserve the empty directories.